+++ src/components/Inputs/index.ts
@@ -1,0 +1,4 @@
+export { default as NumberInput } from './numberInput'
+export { default as RadioInput } from './radioInput'
+export { default as TextArea } from './textArea'
+export { default as TextInput } from './textInput'
