+++ src/components/plugins/Buttons/index.ts
@@ -1,0 +1,3 @@
+export { default as HeadingButton } from './headingButton'
+export { default as ListButton } from './listButton'
+export { default as QuoteButton } from './quoteButton'
